requirements.txt
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.53
langchain_groq==0.1.3
python-dotenv==1.0.1
validators==0.28.1
transformers==4.37.2
unstructured